start to splitting up tests
